requirements_auth.txt
# Backend Authentication Server Requirements
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.1
Authlib==1.3.0
requests==2.31.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
cryptography==41.0.7  # Necessario per Apple Sign In JWT
